==> _notebooks/3. Filter images.ipynb <==
import os
from glob import glob
import rasterio
import numpy as np
import pandas as pd
import cv2
import matplotlib.pyplot as plt

base_dest_dir = '/datasets/rpartsey/satellite/planet/2017-su-au-1_2017-sp-1_2017-sp-2_256x256'
dest_image_dir = os.path.join(base_dest_dir, 'images')
dest_mask_dir = os.path.join(base_dest_dir, 'masks')

images = glob(os.path.join(dest_image_dir, '*.tif'))
masks = glob(os.path.join(dest_mask_dir, '*.tif'))

data = {
    'image': sorted(images),
    'mask': sorted(masks)
}
df = pd.DataFrame(data)

print(df.shape)
df.head()

black_images = []

for _, (image_path, mask_path) in df.iterrows():
    with rasterio.open(image_path) as image_src:
        with rasterio.open(mask_path) as mask_src:
            bands = image_src.read()
            data_mask = bands.sum(axis=0).astype(bool).astype(np.uint8)
            
            mask = mask_src.read()
#             print(data_mask.sum(), (data_mask.sum() < ((256*256) // 2)), not mask.any())
            black_images.append((data_mask.sum() < ((256*256) // 2)) and not mask.any())
    
black_images = np.array(black_images)
print('Number of black images:', np.sum(black_images))

df = df[~black_images]
df.shape

mask_exists = []

for _, (image_path, mask_path) in df.iterrows():
    with rasterio.open(mask_path) as mask_src:
        mask = mask_src.read()
        mask_exists.append(mask.any())
        
mask_exists = np.array(mask_exists)
print('Number of images with burned areas:', mask_exists.sum())

df['mask_exists'] = mask_exists
df['summer-autumn'] = df['image'].map(lambda path: 'summer-autumn' in path)
df['spring'] = df['image'].map(lambda path: 'spring' in path)

print('Numner of spring images:', df['spring'].sum())
print('Numner of summer-autumn images:', df['summer-autumn'].sum())

df.head()

df.to_csv('/datasets/rpartsey/satellite/planet/2017-su-au-1_2017-sp-1_2017-sp-2_256x256_no_black.csv', index=False)

df = pd.read_csv('/datasets/rpartsey/satellite/planet/2017-su-au-1_2017-sp-1_2017-sp-2_256x256_no_black.csv')

df.head()

spring_with_mask = df[df['spring'] & df['mask_exists']]
spring_without_mask = df[df['spring'] & (~df['mask_exists'])]

print(spring_with_mask.shape)
print(spring_without_mask.shape)

summer_autumn_with_mask = df[df['summer-autumn'] & df['mask_exists']]
summer_autumn_without_mask = df[df['summer-autumn'] & (~df['mask_exists'])]

print(summer_autumn_with_mask.shape)
print(summer_autumn_without_mask.shape)

num_samples = 2000

def sample_rows(df, n):
    return df.sample(min(df.shape[0], n))

sp_with_mask  = sample_rows(spring_with_mask, num_samples)
sp_without_mask  = sample_rows(spring_without_mask, num_samples)


su_au_with_mask  = sample_rows(summer_autumn_with_mask, num_samples)
su_au_without_mask  = sample_rows(summer_autumn_without_mask, num_samples)


small_df = pd.concat([sp_with_mask, sp_without_mask, su_au_with_mask, su_au_without_mask])

small_df.shape

small_df.head()

small_df.to_csv('/datasets/rpartsey/satellite/planet/2017-su-au-1_2017-sp-1_2017-sp-2_256x256_no_black_small.csv', index=False)

small_df = pd.read_csv('/datasets/rpartsey/satellite/planet/2017-su-au-1_2017-sp-1_2017-sp-2_256x256_no_black_small.csv')
